Revert "fix: floating point texture size and keyframe deserialization"
This reverts commit 5dfcd0c4af022b16ee7142ec6f4299d44780b619.

+++ src/main/java/de/tomalbrc/bil/file/extra/BbResourcePackGenerator.java
@@ -79,7 +79,7 @@ public class BbResourcePackGenerator {
         }
     }
 
-    public static boolean isMultipleOf(float width, float height) {
+    public static boolean isMultipleOf(int width, int height) {
         return width > 0 && height % width == 0;
     }
 }
